addInt
On the first step of the assignment: I have to create a methd that will take two arrays of "digits" and add them together.

The problem is, is that  I can't figure out what the "i" value should be to get the length right.  The method should accept different lengths of arrays (max of 25).